backend/requirements.txt
# ── Core ─────────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.0
python-multipart>=0.0.20

# ── Database (Supabase client) ───────────────────────────────────
supabase>=2.11.0
httpx>=0.28.0

# ── AI / RAG ─────────────────────────────────────────────────────
llama-index-core>=0.12.0
sentence-transformers>=3.4.0
PyMuPDF>=1.25.0

# ── Thai NLP ─────────────────────────────────────────────────────
pythainlp>=5.0.0

# ── Testing ──────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.25.0
